requirements.micromamba.txt
fire
furl
laspy
einops
opencv
svgwrite
litestar[full]
loguru
#matplotlib
numpy
#open3d
Pillow
plumbum
pymysql
minio
redis-py
pydantic
python-dotenv
python-box[all]
